backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
aspy.yaml==1.1.1          # via pre-commit
astroid==1.6.4            # via pylint
autoflake==1.2
autopep8==1.3.5
backcall==0.1.0           # via ipython
cached-property==1.4.2    # via pre-commit
certifi==2018.4.16        # via requests
cfgv==1.0.0               # via pre-commit
chardet==3.0.4            # via requests
click==6.7                # via flask, pip-tools, safety
decorator==4.3.0          # via ipython, traitlets
dparse==0.4.1             # via safety
first==2.0.1              # via pip-tools
flake8==3.5.0
flask==1.0.2
gunicorn==19.8.1
identify==1.0.18          # via pre-commit
idna==2.6                 # via requests
ipdb==0.11
ipython-genutils==0.2.0   # via traitlets
ipython==6.4.0
isort==4.3.4
itsdangerous==0.24        # via flask
jedi==0.12.0              # via ipython
jinja2==2.10              # via flask
lazy-object-proxy==1.3.1  # via astroid
logzero==1.5.0
markupsafe==1.0           # via jinja2
mccabe==0.6.1             # via flake8, pylint
mypy==0.600
nodeenv==1.3.0            # via pre-commit
packaging==17.1           # via dparse, safety
parso==0.2.1              # via jedi
pep257==0.7.0
pexpect==4.5.0            # via ipython
pickleshare==0.7.4        # via ipython
pip-tools==2.0.2
pre-commit==1.9.0
prompt-toolkit==1.0.15    # via ipython
ptyprocess==0.5.2         # via pexpect
pycodestyle==2.3.1        # via autopep8, flake8
pyflakes==1.6.0           # via autoflake, flake8
pygments==2.2.0           # via ipython
pylint==1.9.1
pyparsing==2.2.0          # via packaging
python-dotenv==0.8.2
pyyaml==3.12              # via aspy.yaml, dparse, pre-commit
requests==2.18.4
rope==0.10.7
safety==1.8.1
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via astroid, cfgv, dparse, packaging, pip-tools, pre-commit, prompt-toolkit, pylint, traitlets
traitlets==4.3.2          # via ipython
typed-ast==1.1.0          # via mypy
urllib3==1.22             # via requests
virtualenv==16.0.0        # via pre-commit
wcwidth==0.1.7            # via prompt-toolkit
werkzeug==0.14.1          # via flask
wrapt==1.10.11            # via astroid
yapf==0.22.0
